+++ paasta_tools/check_chronos_jobs.py
@@ -14,6 +14,7 @@ import service_configuration_lib
 from paasta_tools import chronos_tools
 from paasta_tools import monitoring_tools
 from paasta_tools import utils
+from paasta_tools.chronos_tools import send_event_to_sensu
 
 
 def parse_args():
@@ -43,19 +44,6 @@ def compose_monitoring_overrides_for_service(cluster, service, instance, soa_dir
 def compose_check_name_for_job(service, instance):
     """Compose a sensu check name for a given job"""
     return 'check-chronos-jobs.%s%s%s' % (service, utils.SPACER, instance)
-
-
-def send_event_to_sensu(service, instance, monitoring_overrides, soa_dir, status_code, message):
-    check_name = compose_check_name_for_job(service, instance)
-
-    monitoring_tools.send_event(
-        service=service,
-        check_name=check_name,
-        overrides=monitoring_overrides,
-        status=status_code,
-        output=message,
-        soa_dir=soa_dir,
-    )
 
 
 def last_run_state_for_jobs(jobs):
