+++ paasta_tools/chronos_tools.py
@@ -713,3 +713,21 @@ def find_matching_parent_job(job_name):
         return sorted_jobs[0]['name']
     else:
         return None
+
+
+def compose_check_name_for_job(service, instance):
+    """Compose a sensu check name for a given job"""
+    return 'check-chronos-jobs.%s%s%s' % (service, INTERNAL_SPACER, instance)
+
+
+def send_event_to_sensu(service, instance, monitoring_overrides, soa_dir, status_code, message):
+    check_name = compose_check_name_for_job(service, instance)
+
+    monitoring_tools.send_event(
+        service=service,
+        check_name=check_name,
+        overrides=monitoring_overrides,
+        status=status_code,
+        output=message,
+        soa_dir=soa_dir,
+    )
