+++ tests/test_check_chronos_jobs.py
@@ -29,12 +29,12 @@ def test_compose_monitoring_overrides_for_service(mock_get_runbook, mock_load_ch
 
 def test_compose_check_name_for_job():
     expected_check = 'check-chronos-jobs.myservice.myinstance'
-    assert check_chronos_jobs.compose_check_name_for_job('myservice', 'myinstance') == expected_check
+    assert chronos_tools.compose_check_name_for_job('myservice', 'myinstance') == expected_check
 
 
-@patch('paasta_tools.check_chronos_jobs.monitoring_tools.send_event')
+@patch('paasta_tools.chronos_tools.monitoring_tools.send_event')
 def test_send_event_to_sensu(mock_send_event):
-    check_chronos_jobs.send_event_to_sensu(
+    chronos_tools.send_event_to_sensu(
         service='myservice',
         instance='myinstance',
         monitoring_overrides={},
